feat: update CI workflow to ignore main.rs in coverage report and refactor request handling into serve.rs
