+++ src/main.rs
@@ -1,22 +1,7 @@
-use std::{
-    collections::HashMap,
-    io::{Cursor, Read},
-    path::PathBuf,
-    process::ExitCode,
-    str::FromStr,
-    sync::Arc,
-    time::Duration,
-};
+use std::{io::Read, path::PathBuf, process::ExitCode, sync::Arc, time::Duration};
 
 use anyhow::bail;
-use axum::{
-    Router,
-    body::to_bytes,
-    extract::{Query, Request, State},
-    http::{HeaderName, HeaderValue, Response, status::StatusCode},
-    response::IntoResponse,
-    routing::any,
-};
+use axum::{Router, routing::any};
 use bon::Builder;
 use byte_unit::Byte;
 use clap::{Parser, Subcommand};
@@ -29,8 +14,10 @@ use no_color::is_no_color;
 use rusqlite::Connection;
 use serde_json::{Value, json};
 use tokio::io::{self, AsyncWriteExt as _};
-use tracing::{Instrument, Level, debug, debug_span, error, info};
+use tracing::{Instrument, Level, debug, debug_span, info};
 use tracing_subscriber::{EnvFilter, fmt::format::FmtSpan};
+
+mod serve;
 
 const VERSION: &str = env!("APP_VERSION");
 
@@ -277,8 +264,8 @@ async fn try_main() -> anyhow::Result<()> {
 
 fn build_router(app_state: Arc<AppState>) -> Router {
     Router::new()
-        .route("/{*wildcard}", any(request_handler))
-        .route("/", any(request_handler))
+        .route("/{*wildcard}", any(serve::request_handler))
+        .route("/", any(serve::request_handler))
         .with_state(app_state)
 }
 
@@ -296,129 +283,6 @@ struct AppState {
     timeout: Option<Duration>,
 }
 
-struct AppError(anyhow::Error);
-
-impl IntoResponse for AppError {
-    fn into_response(self) -> axum::response::Response {
-        (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            format!("Something went wrong {}", self.0),
-        )
-            .into_response()
-    }
-}
-
-impl<E> From<E> for AppError
-where
-    E: Into<anyhow::Error>,
-{
-    fn from(err: E) -> Self {
-        Self(err.into())
-    }
-}
-
-async fn try_request_handler(
-    app_state: Arc<AppState>,
-    query: HashMap<String, String>,
-    req: Request,
-) -> anyhow::Result<Response<String>> {
-    let method = json!(req.method().as_str());
-    let path = json!(req.uri().path());
-    let headers = {
-        let mut m = json!({});
-        for (k, v) in req.headers() {
-            m[k.as_str()] = json!(v.to_str()?);
-        }
-        m
-    };
-    let query = json!(query);
-
-    let bytes = to_bytes(
-        req.into_body(),
-        app_state.max_body_size.unwrap_or(10 * 1_024 * 1_024),
-    )
-    .await?;
-    let reader = Cursor::new(bytes);
-
-    let conn = match (app_state.store_path.clone(), app_state.no_store) {
-        (None, None) => Some(Connection::open_in_memory()?),
-        (Some(path), None) => Some(Connection::open(path)?),
-        _ => None,
-    };
-
-    debug!("Evaluating Lua code");
-    let runner = Runner::builder(app_state.source.clone(), reader)
-        .maybe_allow_env(app_state.allow_env.clone())
-        .maybe_default_name(app_state.name.clone())
-        .maybe_http_timeout(app_state.http_timeout)
-        .maybe_store(conn)
-        .maybe_timeout(app_state.timeout)
-        .build()?;
-
-    let request = json!({ "headers": headers, "method": method, "path": path, "query": query });
-    let state = lmb::State::builder()
-        .maybe_state(app_state.state.clone())
-        .request(request)
-        .build();
-    let res = runner.invoke().state(state).call().await?;
-
-    match res.result {
-        Ok(output) => {
-            debug!("Request succeeded: {output}");
-            match output {
-                Value::String(s) => Ok(Response::new(s)),
-                Value::Object(_) => {
-                    let body = output
-                        .pointer("/body")
-                        .map(|v| match v {
-                            Value::String(s) => s.clone(),
-                            _ => v.to_string(),
-                        })
-                        .unwrap_or_default();
-                    let mut res = Response::new(body);
-
-                    let status_code = output.pointer("/status_code").and_then(|v| v.as_u64());
-                    if let Some(status_code) = status_code {
-                        if let Ok(status_code) = u16::try_from(status_code) {
-                            *res.status_mut() = StatusCode::from_u16(status_code)?;
-                        }
-                    }
-
-                    let headers = output.pointer("/headers").and_then(|v| v.as_object());
-                    if let Some(m) = headers {
-                        for (k, v) in m {
-                            if let Some(s) = v.as_str() {
-                                let k = HeaderName::from_str(k.as_str())?;
-                                let v = HeaderValue::from_str(s)?;
-                                res.headers_mut().insert(k, v);
-                            }
-                        }
-                    }
-
-                    Ok(res)
-                }
-                v => Ok(Response::new(v.to_string())),
-            }
-        }
-        Err(err) => {
-            error!("Request failed: {err:?}");
-            Err(err.into())
-        }
-    }
-}
-
-async fn request_handler(
-    State(app_state): State<Arc<AppState>>,
-    Query(query): Query<HashMap<String, String>>,
-    request: Request,
-) -> Result<Response<String>, AppError> {
-    let span = debug_span!("handle_request");
-    let res = try_request_handler(app_state, query, request)
-        .instrument(span)
-        .await?;
-    Ok(res)
-}
-
 #[tokio::main]
 async fn main() -> ExitCode {
     if let Err(e) = try_main().await {
@@ -426,39 +290,4 @@ async fn main() -> ExitCode {
         return ExitCode::FAILURE;
     }
     ExitCode::SUCCESS
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use axum_test::TestServer;
-    use serde_json::{Value, json};
-
-    use crate::{AppState, build_router};
-
-    #[tokio::test]
-    async fn test_serve() {
-        let source = include_str!("./fixtures/serve-echo.lua");
-        let app_state = Arc::new(AppState::builder().source(source).build());
-        let router = build_router(app_state);
-        let server = TestServer::new(router).unwrap();
-
-        let res = server
-            .post("/a/b/c?a=1&b=2")
-            .add_header("i-am", "teapot")
-            .json(&json!({ "foo": 1, "bar": 2 }))
-            .await;
-        assert_eq!(200, res.status_code());
-        assert_eq!(
-            json!({
-                "body": { "foo": 1, "bar": 2 },
-                "headers": { "content-type": "application/json", "i-am": "teapot" },
-                "method": "POST",
-                "path": "/a/b/c",
-                "query": { "a": "1", "b": "2" }
-            }),
-            res.json::<Value>()
-        );
-    }
 }
